💥refactor!: update code to work with synth-css v2
