+++ packages/synth-react/docs/Box.stories.mdx
@@ -6,27 +6,20 @@ import { Box } from '../src'
 
 # Box
 
-The `<Box>` component is a layout primitive used to display a container that supports CSS box model properties.
+The `<Box>` component provides an component-based interface for accessing tokens from your token configuration.
 
 <Preview withToolbar>
   <Story name="intro">
     <Box
-      height={100} 
-      padding="@spacing.3" 
-      style={{ background: 'rgba(255, 255, 255, 0.15)' }}
-      width={260} 
+      backgroundColor="@BeatGig.1"
+      border="input:hover"
+      borderRadius="base"
+      color="@Porcelain"
+      padding="@spacing.3"
+      opacity="modalOverlay"
+      fontSize="heading"
     >
-      260 x 100
-    </Box>
-  </Story>
-</Preview>
-
-You can use the `<Box>` primitive without the default values from your tokens by passing the `withoutDefaults` prop.
-
-<Preview withToolbar>
-  <Story name="withoutDefaults">
-    <Box style={{ background: 'rgba(255, 255, 255, 0.15)' }} withoutDefaults>
-      With no defaults
+      New Box Component
     </Box>
   </Story>
 </Preview>
