--- packages/synth-react/docs/Text.stories.mdx
@@ -1,40 +1,0 @@
-import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks'
-
-import { Text } from '../src'
-
-<Meta title="synth-react|Primitives/Text" component={Text} />
-
-# Text
-
-The `<Text>` component is a layout primitive used to display a container that supports CSS text-related properties.
-
-<Preview withToolbar>
-  <Story name="intro">
-    <div>
-      <Text>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-        tempor incididunt ut labore et dolore magna aliqua.
-      </Text>
-      <br />
-      <Text color="meta">
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-        tempor incididunt ut labore et dolore magna aliqua.
-      </Text>
-    </div>
-  </Story>
-</Preview>
-
-You can use the `<Text>` primitive without the default values from your tokens by passing the `withoutDefaults` prop.
-
-<Preview withToolbar>
-  <Story name="withoutDefaults">
-    <Text withoutDefaults>
-      Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-      tempor incididunt ut labore et dolore magna aliqua.
-    </Text>
-  </Story>
-</Preview>
-
-## Props
-
-<Props of={Text} />
