+++ packages/synth-react/src/Box/index.d.ts
