+++ packages/synth-react/src/Box/index.js
