--- packages/synth-react/src/Primitives/Box/constants.js
@@ -1,88 +1,0 @@
-/**
- * @typedef {string}
- * @constant
- * @default
- * @see {@link https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model/Introduction_to_the_CSS_box_model}
- */
-export const BOX_PROPERTY_BORDER = 'border'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- * @see {@link https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model/Introduction_to_the_CSS_box_model}
- */
-export const BOX_PROPERTY_HEIGHT = 'height'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- * @see {@link https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model/Introduction_to_the_CSS_box_model}
- */
-export const BOX_PROPERTY_MARGIN = 'margin'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- * @see {@link https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model/Introduction_to_the_CSS_box_model}
- */
-export const BOX_PROPERTY_MAX_HEIGHT = 'maxHeight'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- * @see {@link https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model/Introduction_to_the_CSS_box_model}
- */
-export const BOX_PROPERTY_MAX_WIDTH = 'maxWidth'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- * @see {@link https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model/Introduction_to_the_CSS_box_model}
- */
-export const BOX_PROPERTY_MIN_HEIGHT = 'minHeight'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- * @see {@link https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model/Introduction_to_the_CSS_box_model}
- */
-export const BOX_PROPERTY_MIN_WIDTH = 'minWidth'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- * @see {@link https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model/Introduction_to_the_CSS_box_model}
- */
-export const BOX_PROPERTY_PADDING = 'padding'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- * @see {@link https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model/Introduction_to_the_CSS_box_model}
- */
-export const BOX_PROPERTY_WIDTH = 'width'
-
-/**
- * @typedef {string[]}
- * @constant
- * @default
- * @see {@link https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Box_Model/Introduction_to_the_CSS_box_model}
- */
-export const BOX_PROPERTIES = [
-  BOX_PROPERTY_HEIGHT,
-  BOX_PROPERTY_MARGIN,
-  BOX_PROPERTY_MAX_HEIGHT,
-  BOX_PROPERTY_MAX_WIDTH,
-  BOX_PROPERTY_MIN_HEIGHT,
-  BOX_PROPERTY_MIN_WIDTH,
-  BOX_PROPERTY_PADDING,
-  BOX_PROPERTY_WIDTH,
-]
