--- packages/synth-react/src/Primitives/Box/utils/index.js
@@ -1,1 +1,0 @@
-export { default as setBoxProperties } from './setBoxProperties'
