--- packages/synth-react/src/Primitives/Text/constants.js
@@ -1,71 +1,0 @@
-/**
- * @typedef {string}
- * @constant
- * @default
- */
-export const TEXT_PROPERTY_COLOR = 'color'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- */
-export const TEXT_PROPERTY_DECORATION = 'decoration'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- */
-export const TEXT_PROPERTY_FONT = 'font'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- */
-export const TEXT_PROPERTY_LINE_HEIGHT = 'lineHeight'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- */
-export const TEXT_PROPERTY_SIZE = 'size'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- */
-export const TEXT_PROPERTY_SPACING = 'spacing'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- */
-export const TEXT_PROPERTY_TRANSFORM = 'transform'
-
-/**
- * @typedef {string}
- * @constant
- * @default
- */
-export const TEXT_PROPERTY_WEIGHT = 'weight'
-
-/**
- * @typedef {string[]}
- * @constant
- * @default
- */
-export const TEXT_PROPERTIES = [
-  TEXT_PROPERTY_COLOR,
-  TEXT_PROPERTY_DECORATION,
-  TEXT_PROPERTY_FONT,
-  TEXT_PROPERTY_LINE_HEIGHT,
-  TEXT_PROPERTY_SIZE,
-  TEXT_PROPERTY_SPACING,
-  TEXT_PROPERTY_TRANSFORM,
-  TEXT_PROPERTY_WEIGHT,
-]
