--- packages/synth-react/src/Primitives/Text/index.d.ts
@@ -1,6 +1,0 @@
-export {
-  default as Text,
-  TextComponent,
-  TextComponentProps,
-  TextProps,
-} from './Text'
