--- packages/synth-react/src/Primitives/Text/index.js
@@ -1,1 +1,0 @@
-export { default as Text } from './Text'
