--- packages/synth-react/src/Primitives/Text/utils/index.js
@@ -1,1 +1,0 @@
-export { default as setTextProperties } from './setTextProperties'
