--- packages/synth-react/src/Primitives/index.d.ts
@@ -1,2 +1,0 @@
-export { Box, BoxComponent, BoxComponentProps, BoxProps } from './Box'
-export { Text, TextComponent, TextComponentProps, TextProps } from './Text'
