--- packages/synth-react/src/Primitives/index.js
@@ -1,2 +1,0 @@
-export { Box } from './Box'
-export { Text } from './Text'
