+++ packages/synth-react/src/index.js
@@ -1,4 +1,4 @@
-export { Box, Text } from './Primitives'
+export { Box } from './Box'
 export { default as useSynth } from './useSynth'
 export { default as withSynth } from './withSynth'
 export { SynthConsumer, SynthProvider } from './SynthContext'
